Time: 0 ms (100.00%) | Memory: 22.5 MB (69.80%) - LeetSync

+++ 26-remove-duplicates-from-sorted-array/remove-duplicates-from-sorted-array.cpp
@@ -3,8 +3,8 @@ public:
     typedef long long i64;
     int removeDuplicates(vector<int>& nums) {
         int n = nums.size();
-        i64 p = 1;
-        for (i64 i = 0; i < n - 1; i++) {
+        int p = 1;
+        for (int i = 0; i < n - 1; i++) {
             if (nums[i] != nums[i + 1]) {
                 nums[p] = nums[i + 1];
                 p++;
